Update retrogame.c to introduce "debouncing"-array
When playing some high-speed action game, I noticed a delay when moving the joystick. The reason is the kind of "debouncing" ( necessary for low-quality machanical switches): retrogame waits 20ms after GPIO-IRQ until a keystroke is submitted via uinput. This has a severe impact in some action games like "Scrambe". 
So my new logic: When a GPIO-IRQ is received for the first time, no debouncing is done; however a state-change can only occur, when debouncing on this GPIO-Pin times out.
Hope my patch is useful... - enjoy ;-)